[fresh] Reset useMemoCache in response to a FastRefresh run
This PR introduces a new (dev only) debug property on fibers to indicate
that it is being rerendered as a result of a FastRefresh run. This
allows React to clear the cache even where the size (incidentally)
remains constant between renders: for example, if edited code _happens_
to keep the cache size the same but the code is meaningfully different
such that it is no longer correct to reuse the cache.

ghstack-source-id: 2e14ae4bf8aaf9070848138901585d47c2a98e21
Pull Request resolved: https://github.com/facebook/react/pull/30677